IDs int -> int64

+++ printfileinfo.go
@@ -1,7 +1,7 @@
 package model
 
 type PrintfileInfo struct {
-	ProductID           int                `json:"product_id" bson:"product_id"`
+	ProductID           int64              `json:"product_id" bson:"product_id"`
 	AvailablePlacements interface{}        `json:"available_placements" bson:"available_placements"`
 	Printfiles          []Printfile        `json:"printfiles" bson:"printfiles"`
 	VariantPrintfiles   []VariantPrintfile `json:"variant_printfiles" bson:"variant_printfiles"`
